zomato-agent-gemini-langchain/zomato_agent/requirements.txt
langchain_neo4j
python-dotenv
langchain_google_genai
langgraph
langchain-community
faiss-cpu
uvicorn
fastapi
openai
numpy
pandas
pydantic